fix: quotation line-item quantity is now integer, not fractional
Live user testing confirmed quantities are always whole numbers in this
business, superseding the earlier NUMERIC(14,3) design assumption.

- migrations/quotations_schema.sql, init_db_v11.py: quotation_lines.quantity
  and quotation_revision_lines.quantity are now INTEGER (was NUMERIC(14,3)).
  Added idempotent ALTER TABLE ... TYPE INTEGER USING ROUND(quantity)::INTEGER
  statements to migrate the live dev table, rounding any existing fractional
  test rows rather than failing.
- app_pages/quotation_request.py: Quantity number_input is now integer-typed
  (min_value=0, step=1, no format string) so it displays whole numbers and
  rejects decimal input. Prefill logic uses int(...) instead of float(...),
  matching the existing try/except pattern for price/disc.

Verified via information_schema.columns that both columns are now integer
on the live dev DB, migration is idempotent on rerun, imports are clean, and
the full test suite passes (64 passed, 1 skipped, no regressions).

+++ init_db_v11.py
@@ -497,7 +497,7 @@ CREATE TABLE IF NOT EXISTS quotation_lines (
     quotation_id  BIGINT NOT NULL REFERENCES quotation_requests(quotation_id) ON DELETE CASCADE,
     line_no       SMALLINT NOT NULL CHECK (line_no BETWEEN 1 AND 14),
     product_id    TEXT NOT NULL REFERENCES items(product_id),
-    quantity      NUMERIC(14,3) NOT NULL CHECK (quantity > 0),
+    quantity      INTEGER NOT NULL CHECK (quantity > 0),
     unit_price    NUMERIC(14,2) NOT NULL CHECK (unit_price >= 0),
     discount_pct  NUMERIC(5,2)  NOT NULL DEFAULT 0 CHECK (discount_pct BETWEEN 0 AND 100),
     created_at    TIMESTAMPTZ NOT NULL DEFAULT NOW(),
@@ -533,7 +533,7 @@ CREATE TABLE IF NOT EXISTS quotation_revision_lines (
     product_id               TEXT NOT NULL REFERENCES items(product_id),
     article_number_snapshot  TEXT,
     description_snapshot     TEXT,
-    quantity                 NUMERIC(14,3) NOT NULL,
+    quantity                 INTEGER NOT NULL,
     unit_price                NUMERIC(14,2) NOT NULL,
     discount_pct               NUMERIC(5,2) NOT NULL,
     line_total                  NUMERIC(16,2) NOT NULL
